update boundary conditions
git-svn-id: svn+ssh://c4science.ch/source/amie/trunk@1856 64fb8e13-074e-4207-b257-2d34fb7690c6

+++ utilities/granulo.cpp
@@ -44,7 +44,7 @@ Granulo::Granulo()//,double masseInitiale, double densite
 
 }
 
-std::vector <Inclusion *> Granulo::operator()(double rayon_granulat, double pourcentMasseMin, int inclusionNumber , double itzSize, bool verbose)
+std::vector <Inclusion *> Granulo::operator()(double rayon_granulat, double pourcentMasseMin, double inclusionNumber , double itzSize, bool verbose)
 {
     std::vector<Inclusion *> rayon;
