+++ utilities/granulo.h
@@ -48,7 +48,7 @@ public:
 	Granulo() ;
 	virtual ~Granulo() { } ;
 
-	virtual std::vector <Inclusion *> operator()(double , double, int inclusionNumber = 8000, double itzSize = 15e-6, bool verbose = true);
+	virtual std::vector <Inclusion *> operator()(double , double, double inclusionNumber = 8000, double itzSize = 15e-6, bool verbose = true);
 
 	virtual std::vector <EllipsoidalInclusion *> operator()(bool, Point *, double , double, double rfactor = 0.8, int inclusionNumber = 8000, double itzSize = 15e-6, bool verbose = true);
 } ;
